converting to ajax(deleting a post)

+++ controllers/posts_controller.js
@@ -25,7 +25,16 @@ module.exports.create= async function(req,res){
 }
 module.exports.destroy= async function(req,res){
     try{
-    let post= await Post.findById(req.params.id)
+    let post= await Post.findById(req.params.id);
+
+    if(req.xhr){
+        return res.status(200).json({
+            data:{
+                post_id:req.params.id
+            },
+            message:"Post deleted"
+        });
+    }
     if(post.user==req.user.id){
         post.remove();
         await Comment.deleteMany({post:req.params.id});
